Makefile: add target to build static library in addition to DSO
In an effort to incorporate the jitterentropy source with rngd, we had
discussed ways to clone the source file over to the rngd project.
Rather than do that I would like to propose that we instead enable this
library to be built as a static archive, suitable for compile time
linking into the rngd library.  rngd can pull the source code as a
submodule in its git tree, build the static archive and link it in as
needed

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>
Signed-off-by: Stephan Mueller <smueller@chronox.de>