build: Make DSO usabile in development environments
I'd like to move rng-tools away from using a static build of
jitterentropy, and replace it with a dynamic library linkage.  Doing so
makes the distribution of rng-tools a bit easier (people can install
libjitterentropy indepndently, and don't have to go get a second tarball
for building), and it eases security issues (in the event jitterentropy
usage fans out).  To do that though, we need two things:

1) The header files jitterentropy.h and jitterentropy-base-user.h need
to be availble in a standard location

2) The file libjitterentropy.so needs to be available and must point to
the latest version of the library.

This patch enables both of those items.  With this, all the needed files
are in place to allow other projects to detect the presence of the
jitterentropy library via the AC_SEARCH_LIB macro in autoconf.

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>