replace LSFR with SHA-3 256
Replace the non-cryptographic entropy compressing operation of an LFSR
with SHA3-256. Besides using a cryptographic hash operation for
compressing entropy, the entropy analysis for SP800-90B is expected to
be significantly easier.

Due to the large amount of operations of a SHA3-256 instead of an LFSR,
the code now can be compiled with optimizations like -O2. All entropy
assessments producing the following numbers are generated with -O2.

The maximum number of SHA3-256 repetitions is between 2^0 + 1 and 2^3 +
1.

Using 4-bit time windows, the following SP800-90B values are obtained on
user space Intel Core i7 8565U:

- entropy rate of 1,000,000 samples:

	* single: 2,78 bits

	* var: 3,56 bits

- entropy rate of reboot tests column-wise:

	* single: 2.15 bits

	* var: 2.07 bits

- entropy rate of reboot tests row-wise:

	* single: 2.11 bits

	* var: 2.03 bits

The performance of the generation is about 35 kBytes/s.

The user visible change is that the block size of the Jitter RNG is
equal to the SHA3-256 digest size.

The SHA3-256 implementation is ACVP tested with the ACVP certificate
A770 (see [1]).

[1] https://csrc.nist.gov/projects/cryptographic-algorithm-validation-program/details?product=13214

Signed-off-by: Stephan Mueller <smueller@chronox.de>