Provide internal timer
In case the underlying hardware does not provide a suitable timer, such
as when the time stamp is too coarse, the Jitter RNG may revert to an
internal high-resolution timer.

This internal timer is implemented by having a counting loop executing
in a separate thread and obtain the counter value when a time stamp
value is requested (i.e. after the SHA-3 / memory access operations).

The thread is spawned anew each time a new random number shall be
generated.

The patch also fixes the X86 time gathering macros.

Signed-off-by: Stephan Mueller <smueller@chronox.de>