correct verification of delta_sum
`delta_sum` is a running sum of the absolute value of second-order-delta
values. This sum is likely to be large in a correctly functioning system,
and is never normalized by dividing it by the number of iterations.

The current code can never actually evaluate as true; delta_sum evaluates
to 0 iff every delta is a fixed value, but in this case the source would
have tested as "stuck" earlier, and exited as a result. In order to be 1,
it would have to be "stuck" for all values other than 1, which would
again necessarily have triggered an early exit due to it being "stuck".
I suspect that this should instead be compared against
`JENT_POWERUP_TESTLOOPCOUNT` or the delta_sum value should be
normalized by dividing it by this value prior to this test.

Suggested-by: "Joshua E. Hill"
Signed-off-by: Stephan Mueller <smueller@chronox.de>