Remove variating SHA3 invocation
Some reviewers suggested that the variations in the numbers of SHA3
invocations performed for one time delta value makes the entropy
assessment more complicated than necessary. Moreover it is considered
that the variations are not adding entropy and thus the baseline of one
SHA3 invocation operation is to be used for entropy assessment.

To ease the mathematical review, remove the variations in the SHA3
invocation. To offset the change, the minimum oversampling rate is set
to 3 which implies that one time delta is considered to have 1/3rd bit
of entropy. Note, the value of the minimum OSR of 3 applies is
considered to be sufficient when compiling the Jitter RNG without
optimizations. The next patch provides the invocation without
optimizations.

Suggested-by: Joshua E. Hill
Suggested-by: Kurt Roeckx
Signed-off-by: Stephan Mueller <smueller@chronox.de>