The existing RCT and APT cutoffs weren’t calculated correctly; the RCT was rounded incorrectly, and the prior APT cutoff didn’t account for a difference in the meaning of apt_count (in this library) and B (in 90B).
Additionally, the library generally assumes a min entropy lower bound
entropy of 1/osr, but the APT cutoff was fixed and based on the old
assumption of 1 bit of min entropy per symbol (the estimate has since
been reduced to 1/3 by default). This pull request makes the cutoff
dependent on the choice of osr (and thus the presumed entropy level),
which is consistent with how the RCT cutoff is treated.

Finally, the APT test that was implemented wasn’t quite what was
specified in 90B (due to reuse of the last symbol of a window as the
first symbol of the next window).