Support for external thread handler
The Jitter RNG uses threading support if it operates in timer-less mode
(i.e. when it detects that the underlying hardware timer is too coarse).
In this case, the Jitter RNG spawns a thread that acts as a cycle
counter.

To allow a caller to manage that thread, the Jitter RNG offers an API to
set a thread handler.

Signed-off-by: Stephan Mueller <smueller@chronox.de>