Fix outfile based on filepath
In most cases, $filepath is already a full path, so it is wrong to prepend
$RESULTS_DIR to it.  This happens to work with some relative paths, but
doesn't work for most relative or absolute paths.

Note: there is a case in validate-runtime/processdata.sh where $filepath
is only the base name (line 160). In that case, $filepath is not an accurate
variable name.  I did not fix that case.

Signed-off-by: Stephan Mueller <smueller@chronox.de>