test: What to do in case of insufficient entropy?
With the recent changes, users started to complain that insufficient
entropy is present. The entropy rate can be increased by enlarging the
memory buffer accessed by the Jitter RNG. The larger the memory buffer,
the more entropy is harvested by the Jitter RNG.

To find a good balance between a decent memory buffer size and
sufficient entropy, an analyzer tool is provided that collects the
raw entropy rate for many different memory settings. In addition,
runtime and restart analysis tools are provided to process the collected
data. In the end a list of entropy rates for the different memory
settings is provided. This allows the user to find the proper memory
configuration for this environment.

The change implies that the memory settings are turned into a compile
time option that can be set with CFLAGS. Thus, jitterentropy.h is
updated allowing to set such compiler flags.

The readme is updated explaining how to start the test and how to
analyze and apply the results.

Signed-off-by: Stephan Mueller <smueller@chronox.de>