JENT: Increase memory buffer used for memory access loop
The Jitter RNG accesses the memory to add more CPU-based uncertainty due
to wait states when accessing memory locations. Small memory buffers may
be completely covered with the L1 cache which implies that hardly any
variations are added. Thus, the memory should be larger than this cache.

The new memory definition implies that 512 kBytes of memory is
allocated. Measurements on a systems have shown that this value
increases the available entropy significantly, even on systems that used
to have insufficient entropy. For example, on a RISC-V system which
before the patch had some entropy rate of around 0.01 bits of entropy
per time delta now have 0.5 bits of entropy. This implies that before
the change the Jitter RNG was not collecting sufficent entropy. After
the change, it collects sufficient entropy as the value is larger than
1/3 (the heuristically applied entropy rate by the Jitter RNG).

The analysis was performed using the newly added analyze_options.sh
scripts.

Signed-off-by: Stephan Mueller <smueller@chronox.de>