JENT: start internal timer thread only on 2 or more CPUs
By adding the service function of jent_ncpu(), the number of CPUs in the
underlying system is detected. Only if 2 or more CPUs are found, the
internal timer thread is enabled.

This change also contains a fix that enabled the timer thread only when
the internal timer is truly needed.

Suggested-by: Vladis Dronov
Signed-off-by: Stephan Mueller <smueller@chronox.de>