JENT: Fix APT
The library generally assumes a min entropy lower bound entropy of
1/osr, but the APT cutoff was fixed and based on the old assumption
of 1 bit of min entropy per raw sample (the estimate has since been
reduced to 1/3 by default). This pull request makes the cutoff
dependent on the choice of osr (and thus the presumed entropy level),
which is consistent with how the RCT cutoff is treated.

Finally, the APT test that was implemented wasn’t quite what was
specified in 90B (due to reuse of the last symbol of a window as the
first symbol of the next window). I doubt this difference had any real
impact, but it does impact SP 800-90B validation.

Sighed-off-by: Joshua E. Hill
Signed-off-by: Stephan Mueller <smueller@chronox.de>