JENT: Force inline of code
The function to insert the data into the GCD data array should be done
such that in the best case it has no measurable impact on the overall
execution that analyzes the system behavior. Thus, the code sould be
inlined. Yet, the inline marker is not honored with -O0.

Thus, the function is implemented as a macro that ensures it is inlined
during compilation.

Reported-by: Joshua E. Hill
Signed-off-by: Stephan Mueller <smueller@chronox.de>