JENT: add jent_read_entropy_safe
This is a service function to jent_read_entropy() with the difference
that it automatically re-allocates the entropy collector if a health
test failure is observed. Before reallocation, a new power-on health test
is performed. The allocation of the new entropy collector automatically
increases the OSR by one. This is done based on the idea that a health
test failure indicates that the assumed entropy rate is too high.

Note the function returns with an health test error if the OSR is
getting too large. If an error is returned by this function, the Jitter RNG
is not safe to be used on the current system.

To support the function, the flags used to initialize the entropy
collector with are stored in the entropy collector, too.

Signed-off-by: Stephan Mueller <smueller@chronox.de>