JENT: remove test for variations < 100
During power-up a test was executed that required the number of
variations in the last two digits to be larger than 90%. Considering
the presence of the GCD calculation, this test is now considered to be
supoerseeded by it. The GCD test requires that the GCD is below 100. In
addition, a test is present that ensures that that variations do exist
in all measurements.

This allows timer variations where a large number of values change above
100 or even 1000 and yet the noise source is still considered
appropriate. Thus, the simple test mandating that 90% of all variations
must show changes where the last two digits change is removed.

One particular example is a RISC-V U74 CPU where the majority of the
changes happen above 10000 with yet sufficient entropy. The now removed
test fails this CPU which is wrong.

Signed-off-by: Stephan Mueller <smueller@chronox.de>