JENT: properly increase osr in jent_read_entropy_safe
Using osr++ as a parameter implies that the incement is performed after
the parameter is sent to the invoked function. As we want the increment
to take place before, the change now increments the variable when it is
initialized.

This also ensures that the validation of the maximum value is now
correct. The maximum value of 20 is taken from the available lag
predictor OSR thresholds. Therefore any value larger than 20 is not
relevant.

Reported-by: Joshua E. Hill
Signed-off-by: Stephan Mueller <smueller@chronox.de>