Add a possibility to use hardware and OS timers
Since v3.0 jitterentropy uses RDTSC on an x86_64 and clock_gettime() on
other arches unconditionally. This has led to a regression versus v2.2
in a certain corner case where RDTSC timer does not pass certain checks.

This change suggests a possibility to support both hardware (RDTSC on
an x86_64, MFTB on Power, etc) and OS (clock_gettime()) timers and an
ability to choose which timer to use. This patch also introduces a kind
of a framework to easily add more timer types.

Link: https://bugzilla.redhat.com/show_bug.cgi?id=1974132#c17

Signed-off-by: Vladis Dronov <vdronov@redhat.com>