JENT: add jent_entropy_init_ex
Prior to this API call, the jent_entropy_init function would always use
the default OSR and flags values. If the caller, however, wants to use
different flags/OSR values, they were not considered in the power-on
self test. This would imply that the power-on self test is not testing
the exact configuration used at runtime.

The jent_entropy_init_ex API now allows the caller to specify the
OSR/flags field with the idea that they should be identical to the one
used with jent_entropy_collector_alloc.

The presence of this new API call now allows to add a safeguard to
jent_entropy_collector_alloc_internal enforcing the execution of the
self tests if they have not been run yet. This is implemented with a
global variable. Considering that the Jitter RNG library is intended to
be re-entrant and thus thread-safe, the use of the global variable here
is considered safe nonetheless because this flag applies to the
underlying system (i.e. if the self test passed once, it is expected to
pass again). Therefore, the use of a global variable here is not
expected to violate the basic concept of the Jitter RNG library.

Signed-off-by: Stephan Mueller <smueller@chronox.de>