JENT: Fix potentially overflowing expression
Fix potentially overflowing expression "1U << min" with type "unsigned int"
(32 bits, unsigned) which is evaluated using 32-bit arithmetic, and then used
in a context that expects an expression of type "uint64_t" (64 bits, unsigned).

This was reported by a covscan (Coverity Scan static analysis tool).

The use of UINT64_C(1) is considered the appropriate solution as that forces
the literal value to be of the correct type.

In addition, a compile-time check is added to make sure the input
parameters to jent_loop_shuffle cannot overflow the bit shift.

Suggested-by: Vladis Dronov
Suggested-by: Joshua Hill
Signed-off-by: Stephan Mueller <smueller@chronox.de>