JENT: SHA-3 - simplify memory management
The allocation of memory on stack is simplified as we have one fully
defined data structure.

Note, the old code was copied from an implementation of different SHA
types where the context data structure varies. For that case, the more
generic old case was applicable. Yet, for the Jitter RNG we only have
SHA3-256 and thus only need to consider this context.

The changes are verified with ACVP.

Reported-by: johanmw
Signed-off-by: Stephan Mueller <smueller@chronox.de>