JENT: random memaccess - determine cache size at runtime
It is possible to obtain the cache size at runtime on some platforms.
This is used in jent_cache_size_roundup which calculates the next higher
integer value that is a power of 2 from the sum of the available caches.

This size is used to allocate the memory for random memory access.

If the cache information is not available, a default memory size is
used. This is currently defined to be 1<<17 based on measurements on a
low-entropy RISC system. This RISC system is measured with the following
memory accesses:

Number of bits  min entropy
10       0.274937
11       0.290911
12       0.233623
13       0.208554
14       0.252896
15       0.358300
16       0.487093
17       0.523763
18       0.512144
19       0.495936
20       0.502976
21       1.131450
22       1.662580
23       1.731049
24       1.381146

The entropy value for 17 bits is considered sufficient for the OSR of 3.
On other systems, we get much more entropy and thus the bit value is
considered to be appropriate as well.

Signed-off-by: Stephan Mueller <smueller@chronox.de>