JENT: add static marker
For some users like the libgcrypt, the functions exported by the Jitter
RNG are used as an internal API that should not be publicly exported.
Thus, all API calls should be equipped with a marker allowing the
environment to declare whether the API call is public or marked as
static. Add the missing marker to newly added API calls.

Signed-off-by: Stephan Mueller <smueller@chronox.de>