JENT: silence clang-analyzer
The clang-analyzer reports the following issue:

src/jitterentropy-noise.c:213:15: warning: The left expression of the compound assignment is an uninitialized value. The computed value will also be garbage [core.uninitialized.Assign]
                curState[i] ^= ec->data[i];

After looking at the pointer values of the buffer, I do not see any
problem with it. But in order to silence the analyzer, instead of a
cast, a union is now used that does the exact same as a type cast. This
silences the analyzer without changing the logic.

Signed-off-by: Stephan Mueller <smueller@chronox.de>