Add a possibility to use hardware and OS timers
Since v3.0 jitterentropy uses RDTSC on an x86_64 and clock_gettime() on
other arches unconditionally. This has led to a regression versus v2.2
in a certain corner case where RDTSC timer does not pass certain checks.

This change suggests a possibility to support both hardware (RDTSC on
an x86_64, MFTB on Power, etc) and OS (clock_gettime()) timers and an
ability to choose which timer to use.

Notes:

A presense of CPUID command has to be checked in a possible future port
to i386, otherwise we can get #UD on older CPUs. See __get_cpuid_max()
for detecting CPUID presense on __i386__:
https://sites.uclouvain.be/SystInfo/usr/include/cpuid.h.html

A single state variable describing which timer should be used for
an allocated entropy context would uncomplicate the runtime processing.
This does not seem possbile now without breaking an API, but should
be probable in the next major version.

A execution flow error was fixed in jent_entropy_init_ex().

An argument order was fixed in jent_notime_enable().

Link: https://bugzilla.redhat.com/show_bug.cgi?id=1974132#c17
Signed-off-by: Vladis Dronov <vdronov@redhat.com>