JENT: Add option to set maximum memory
With the addition to derive the size of the memory access memory from
the size of the CPU data cache, the amount of memory allocated by the
Jitter RNG may become very large. This is good as it brings more
entropy, but it is equally not good as it may use precious resources.

With this change, the maximum amount of memory used per default is
reverted back to the previous level defined by JENT_MEMORY_SIZE. Yet,
the caller may now specify the maximum amount of memory it allows the
Jitter RNG to allocate with the flags JENT_MAX_MEMSIZE_*. On systems
with memory pressure but ample entropy, a lower value might be in order.
On systems with low entropy, a larger memory value may be configured
during allocation.

No matter how large the caller's memory size is, the absolute maximum
amount of memory the Jitter RNG uses is defined by the CPU data cache
size rounded up to the nearest power of 2.

In addition, the change alters jent_read_entropy_safe such that it
increases the used memory in case insufficient entropy is detected by
one level. Yet, the automatic increase will cause the Jitter RNG to
allocate up to the CPU data cache size rounded up to the nearest power
of 2 for the memory access operation.

The tests are updated to allow testing for the memory size using command
line switches allowing the user to set the maximum memory.

Signed-off-by: Stephan Mueller <smueller@chronox.de>