Fix an execution flow and an arguments order
A value of "ret" is always 0 after the first "if (ret)" in
jent_entropy_init_ex(). This means jent_time_entropy_init()
in the third "if (ret && ..." is not called if
JENT_FORCE_INTERNAL_TIMER is set in "flags" and so
"ret = jent_time_entropy_init()" in the second "if" is skipped.
Fix this by assigning an initial non-zero value to "ret".

Also fix an arguments ordering in jent_notime_enable().

Signed-off-by: Vladis Dronov <vdronov@redhat.com>
Signed-off-by: Stephan Mueller <smueller@chronox.de>