JENT: Modify the integration of SHA-3
The entropy pool is now a SHA-3 state. This state is filled with the
time delta containing entropy and auxiliary data that does not contain
entropy using a SHA update operation. When random data shall be produced
from the SHA-3 conditioning operation, a SHA-3 final of the entropy pool
state is obtained followed by an immediate re-initialization of the
SHA-3 state for subsequent operations.

The auxiliary data is calculated by a SHA-3 hashing of some varying
state data. The time delta that contains entropy is measured over the
SHA-3 hasing operation of the auxiliary data. This satisfies FIPS 140-3
IG D.K resolutions 4, 6, and 8 which defines that entropy that is in
an entropy pool shall not be hashed again when claiming that the data
shall still have entropy.

Mathematically speaking, the random number generated by the Jitter RNG
is:

aux_t = SHA-3(Jitter RNG state data)

Jitter RNG block = SHA-3(time_i || aux_i || time_(i-1) || aux_(i-1) ||
			 time_(i-767) || aux_(i-767))

when assuming that the OSR = 3, i.e. the default value.

Suggested-by: Marcos Portnoi
Signed-off-by: Stephan Mueller <smueller@chronox.de>