JENT: link Jitter RNG random numbers
Prior to using the SHA message digest context as a state for the Jitter
RNG, all random numbers produced by the Jitter RNG were intrinsically
linked with each other because the internal state of the Jitter RNG was
never cleaned.

With the sha3_final in jent_read_random_block, this is not longer true.
To re-establish the previous behavior of linking the different random
number blocks, re-insert the message digest calculated from the current
hash state into the newly initialized state that is to be used for the
next generation round. The current digest is not considered to have
entropy, but only to "stir" the hash state a bit.

Signed-off-by: Stephan Mueller <smueller@chronox.de>