JENT: speed up enhanced backtracking support
The enhanced backtracking support so far was based on generating a new
random number by collecting new fresh entropy. That is very time
consuming.

A different approach provided with this patch is to simply calculate the
message digest of the current state and use it as a starting point for a
new state. This way, any residual information (a) is protected by the
strength of the message digest and (b) the old information is still used
to stir the new hash state for future random numbers.

As the calculation of a message digest is much faster than generation a
complete new set of entropy data, the enhanced backtracking support
received a performance boost.

Signed-off-by: Stephan Mueller <smueller@chronox.de>