obtain time for loop shuffle operation
In the currently disabled code to shuffle the loop counter, resurrect
the gathering of the time stamp as a basis for the shuffle operation.
This invocation was lost for 3.4.0.

In addition, perform the shuffle calculation only for the size of the
time stamp value, i.e. 64 bits. More operations increase the time
consumed in the shuffle calculation without changing the value.

Reported-by: Joshua E. Hill
Signed-off-by: Stephan Mueller <smueller@chronox.de>