Remove bi-modal timing behavior of conditioner
As identified by Josh Hill, the execution time of the conditioner exhibits
a bi-modal behavior. This is based in the nature of the used SHA3-256
which only performs a Keccak operation if sufficient data was added to
the hash stack. This implies that not in all insertions of a timing event
a Keccak operation was performed.

By using a buffer that is exactly SHA-3-256-rate-size to hold all data
to be inserted into the entropy pool, it is guaranteed that for each
insertion of data always one Keccak operation is performed.

Signed-off-by: Josh Hill
Signed-off-by: Stephan Mueller <smueller@chronox.de>