Add support for the JENT_CACHE_SHIFT_BITS macro, which allows the developer to request a larger memory region. In this PR, the desired memory region size is now effectively 2^JENT_CACHE_SHIFT_BITS * jent_cache_size_roundup(). The default behavior in this PR is to set JENT_CACHE_SHIFT_BITS to 0 (which is equivalent to the current baseline behavior), but this can be overridden during compilation.
In the current baseline code, there is no mechanism for requesting a memory
region significantly larger than the available cache size. In some cases
the behavior of the JEnt library improves when using a larger memory region,
so there should be some way to induce such a selection.

In all cases, the size of the allocated memory region is mediated by the passed
in JENT_MAX_MEMSIZE_* flag or JENT_MEMORY_SIZE (if no such flag is provided).