Improvements for userspace hashtime collection.
The existing tool always produces two columns of textual data, even
in the instance where `JENT_CONF_DISABLE_LOOP_SHUFFLE` is defined
(whence, the meaning of the two columns is the same).
This macro is the default setting, so this behavior effectively
doubles the collection time in most situations, without any particular
benefit.

After application of this PR, `jitterentropy-hashtime` produces only
one column of data when `JENT_CONF_DISABLE_LOOP_SHUFFLE` is defined.

This PR also adds the ability to output the data in binary format
(uint64_t values in machine format) when only one column would be
output which can be useful in many testing situations.

Finally, as the size of the allocated memory region can be somewhat
obscure, the `jitterentropy-hashtime` tool now outputs the size of the
memory region being used when such information is available (i.e.,
when `JENT_RANDOM_MEMACCESS` is defined, which is the case by default).

My testing seems to indicate that this change doesn't break the testing
infrastructure (the various "single" results just contain an
error message from ea_non_iid rather than results).

In both the current baseline behavior and in the behavior after
applying this PR, the labeling of results is somewhat misleading, as
when  `JENT_CONF_DISABLE_LOOP_SHUFFLE` is defined, all results
reflect the "single" (i.e., no loop shuffle) behavior.