Add a tool "jitterentropy-osr" that allows selection of a maximal osr bound that would be acceptable under the constraint that each 256-bit block must be produced within a particular time period.
For example, if one wanted to use the average runtime across 100 rounds
of output and targeted an upper-bound of 250ms, one might run the command:

```
jitterentropy-osr 100 0.25 --force-fips --disable-internal-timer
```

and the sample output would be something like this:
```
Minimum osr=1 invariant verified.
Maximum osr bound 64 found.
Desired osr bound is in [32, 48]
Desired osr bound is in [40, 48]
Desired osr bound is in [44, 48]
Desired osr bound is in [46, 48]
Desired osr bound is in [47, 48]
osr bound is 47
```

Signed-off-by: Stephan Mueller <smueller@chronox.de>