Kernel Tester: Calculate time delta
The Linux kernel returns the time stamps to user space. This implies that
the recording used to be subject to post-processing to obtain the time
deltas. Now, the tool is updated to generate the delta already during
recording to allow the data to be directly used by the analysis tools.
Therefore, the data is now of the same format as the user space recording.

In addition, the data format of the raw time stamps changed from 32 to 64
bits.

Signed-off-by: Stephan Mueller <smueller@chronox.de>