AIS 20/31 v3.0 NTG.1 support
AIS 20/31 v3.0 requires for NTG.1 for initialization to obtain entropy
from two different noise source each delivering 240 bits of entropy
independently.

The approach is to treat the memory access and the SHA3 timing
as independent noise sources. Therefore, the initialization first
obtains 256 bits of entropy from the memory access alone followed by
256 bits of entropy from the SHA3 operation. All data is inserted into
the entropy pool. Only after the completion of this operation, a 256
bit message digest is generated from the entropy pool.

After this initialization, the Jitter RNG behaves regularly.

The NTG.1 behavior is only enabled if the flag JENT_NTG1 is set during
initialization. If the flag is not set as well as after the NTG.1
initialization, the Jitter RNG shall have no difference to the common
approach. Also, the common approach does not show any changes in the
characteristics.

To achieve the NTG.1 implementation but having no effect on the common
case, the following changes are applied:

- jent_hash_time is butchered into two independent functions:
  jent_hash_loop implements the hash loop and jent_hash_insert which
  inserts any data into the entropy pool. Although two functions are
  present now, both implement the identical functionality of the original
  function.

- jent_measure_jitter is changed to invoke the two new functions instead
  of the old one. This change does not introduce any functional
  change of any kind.

- addition of the jent_measure_jitter_ntg1_* which independently
  trigger the memory access and the hash loop noise source. I.e. these
  functions measure the execution time over only the memory access and
  the hash loop. These functions invoke the stuck/health tests and the
  insertion of the data into the entropy pool identically to
  jent_measure_jitter.

- jent_random_data is updated to invoke jent_measure_jitter_ntg1_* if
  NTG.1 is selected. This invocation is only performed during
  startup - at runtime or without NTG.1 the already known
  jent_measure_jitter function is used which implies that the common
  entropy collection is used.

- addition of the flag JENT_NTG1 which is required during initialization
  by the caller to trigger the NTG.1 behavior.

This change functionally works, but the calibration and entropy assessment
on different systems is not yet performed. Changes are expected after the
conclusion of the detailed analysis.

Signed-off-by: Stephan Mueller <smueller@chronox.de>