NTG.1: Increase hash/mem loop count when used as individual noise source
During startup of an NTG.1 compliant behavior, the hash loop as well as
the memory access loop serve as individual entropy sources. They must
deliver sufficient entropy to reach the threshold of 240 bits of entropy.

To keep the overall user-visible operation simple, the same OSR value
is used for the common operation (hash/memory loop are used in unison)
as well as the NTG.1 startup operation (individual hash / memory loop
noise source). To offset the respectively missing noise source in the
NTG.1 startup when delivering equal or more than the heuristically
implied entropy value of 1/OSR, increase the memory / hash loop count
by three times and yet keep the OSR unchanged.

With this modification, the NTG.1 behavior was tested successfully
on border-line systems as follows. For the following table, the
entropy rate from the SP800-90B tool of the common behavior, the
individual hash loop and the individual memory access loop are measured.
The following systems are borderline in respect to the following
behavior:

- Apple M4 Max: The memory access seems to be so streamlined that
there is little entropy found in it compared to other systems.

- Spacemit X60 RISC-V system: The CPU instruction variations are very
limited due to the fact that this CPU is very small with very
limited internal components that are on different clock domains.

To put it into perspective, additional measurements are provided on
a system with comparatively large entropy: an ARMv7 32 bit system
and an Intel Alder Lake Core Ultra 7 155H.

            | Common |  Mem Acc Loop | Hash Loop
------------+--------+---------------+--------------
M4 Max      |  1.18  |      0.82     | 1.58
RISC-V      |  5.90  |      3.02     | 0.44
ARMv7       |  7.27  |      7.18     | 3.69
Ultra 7 155 |  5.69  |      3.37     | 5.99

All systems, including the border-line systems show sufficient
entropy for the Jitter RNG as they are more than 1/3 bits of
entropy (using the default OSR of 3 implying a heuristic entropy
rate of 1/3).

Furthermore, the generation of random bits did not trigger health
test errors.

Side note: The Jitter RNG in the common behavior remains unchanged
after this patch.

Signed-off-by: Stephan Mueller <smueller@chronox.de>