NTG.1: Add raw entropy measurement ability
The existing tools for statistically measuring the entropy rate using
the SP800-90B tool set is updated to remove the differentiation between
variating and single loop behavior of the noise source due to this
functionality being removed from the Jitter RNG (see patch
"Remove all code when JENT_CONF_DISABLE_LOOP_SHUFFLE is not set").

In addition, the test tool set is updated to accommodate the NTG.1
testing with the following changes:

* Add a recording_userspace/invoke_testing_ntg1.sh tool which
obtains the raw entropy data for the common entropy source behavior,
the hash-loop only behavior and the memory-access-loop-only behavior.
This allows the calculation of the statistical entropy rate for all
three variants independently.

* Add a validation-runtime/processdata_ntg1.sh too to calculate
the statistical entropy rate using the SP800-90B tool for each
of the aforementioned three noise source data sets. All three
statistical data sets allow the validation of the entropy rate
for the NTG.1 startup requirement as well as common runtime
behavior.

* Add options to jitterentropy-hashtime to add more command line
options for allowing the caller to set the different flags when
allocating the Jitter RNG. This approach shall allow the caller
to precisely set the Jitter RNG behavior to be measured.

When testing the common Jitter RNG use case, the command line options
that were applicable before still work unchanged.

Signed-off-by: Stephan Mueller <smueller@chronox.de>